Build overhaul (#246)
* Combine the publish workflows

* Upgrade build system and fix compile warnings

* Mark python 3.13 as supported

* First pytest pass

Currently, all tests except for enspara/test/test_smfret.py pass;

* More packaging fixes

* Fix test_dye_emission

This should have no impact on actual functionality. I assume a change in the numpy RNG caused this test to fail.

* Update pyproject.toml with more info

* "It's an entirely new test suite I promise" - me lying to GitHub

* Some CI fixes

* Justin says this is probably fine?

* Undisable the pytest MPI stuff

* Well that would do it

* Fix test failures on darwin

Apparently there's an error on the order of 10^-7 but only on darwin. I'm assuming this is floating point nonsense

* Robust test_rmsd_khybrid_mpi_basic

The old version depended on fragile behavior of mpich, this new one performs basically the same tests but works much better

* Add  workflow_dispatch to publish

* remove outdated load in smFRET code

---------

Co-authored-by: Justin Miller <66532328+Justin-J-Miller@users.noreply.github.com>